feat : 구글, 네이버 로그인 redirect uri 설정 및 일부 아이콘 변경
